jasper Reports added.

+++ src/main/java/com/example/Studentdatamanagement/service/StudentService.java
@@ -1,6 +1,8 @@
 package com.example.Studentdatamanagement.service;
 
 import com.example.Studentdatamanagement.Repo.StudentRepository;
+import com.example.Studentdatamanagement.Response.AddressResponse;
+import com.example.Studentdatamanagement.Response.StudentResponse;
 import com.example.Studentdatamanagement.entity.Student;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -8,6 +10,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.reactive.function.client.WebClient;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class StudentService {
@@ -20,6 +23,18 @@ public class StudentService {
     @Autowired
     private WebClient webClient;
 
+    public StudentResponse getStudentById(int id) {
+
+        Optional<Student> student = repository.findById(id);
+       StudentResponse studentResponse = mapper.map(student, StudentResponse.class);
+
+        // Using WebClient
+        AddressResponse addressResponse = webClient.get().uri("/students/" + id).retrieve().bodyToMono(AddressResponse.class).block();
+        studentResponse.setAddressResponse(addressResponse);
+        return studentResponse;
+    }
+
+
     public Student saveStudent(Student student){
       return repository.save(student);
     }
@@ -31,20 +46,15 @@ public class StudentService {
         return repository.findAll();
     }
 
-    public Student getStudentById(Long id){
-        return repository.findById(id).orElse(null);
-    }
+//    public Student getStudentById(Long id){
+//        return repository.findById(id).orElse(null);
+//    }
 
-
-
-
-    public Student getStudentByName(String name){
-        return repository.findByName(name);
-    }
-
-    public String deleteStudent(Long id){
+    public void deleteStudent(int id) {
+        if (!repository.existsById(id)) {
+            throw new StudentNotFoundException("Student not found with ID: " + id);
+        }
         repository.deleteById(id);
-        return "Student removed: " +id;
     }
 
     public Student UpdateStudent(Student student){
@@ -56,6 +66,12 @@ public class StudentService {
          return repository.save(exsistingStudent);
     }
 
+
+    public class StudentNotFoundException extends RuntimeException {
+        public StudentNotFoundException(String message) {
+            super(message);
+        }
+    }
 
 
 }
